Set target location of rabbitmq_data share to /var/lib/rabbitmq/mnesia
In order to allow the ussage of a cifs share for the rabbitmq_data
directory we need to append mnesia to the target location. As the
current path /var/lib/rabbitmq hosts the .erlang.cookies file, that
requires strict posix permissions, which won't work on cifs out of the box.

See also: https://github.com/docker-library/rabbitmq/issues/171#issuecomment-316302131